refactor: add compression strategy
